App/flask/requirements.txt
gunicorn==19.9.0
category_encoders==1.3.0
joblib==0.13.2
numpy==1.15.4
pydotplus==2.0.2
PyMySQL==0.9.3
Werkzeug==0.14.1
pandas==0.23.4
Flask==1.0.2
flask_paginate==0.5.2
matplotlib==3.0.2
scikit_learn==0.20.3
sqlalchemy
mysql-connector-python
